first_project/requirements.txt
Django==5.1.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-debug-toolbar==6.2.0
django-unfold==0.40.0
Pillow==10.4.0
celery==5.3.6
redis==5.0.1
